Rename tables to joining_type_tables
